Move lambda.execution.sam to lambda.execution.local package
* Rename classes that aren't tied to Sam to LocalLambda to match RemoteLambda
* Move utilities and common code not tied to locally executing a lambda from the package
