+++ jetbrains-core/it/software/aws/toolkits/jetbrains/services/lambda/execution/local/SamTestUtils.kt
@@ -1,7 +1,7 @@
 // Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package software.aws.toolkits.jetbrains.services.lambda.execution.sam
+package software.aws.toolkits.jetbrains.services.lambda.execution.local
 
 import com.intellij.execution.ExecutorRegistry
 import com.intellij.execution.Output
