+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/lambda/execution/local/LocalLamdaRunConfiguration.kt
@@ -1,7 +1,7 @@
 // Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package software.aws.toolkits.jetbrains.services.lambda.execution.sam
+package software.aws.toolkits.jetbrains.services.lambda.execution.local
 
 import com.intellij.execution.ExecutionException
 import com.intellij.execution.Executor
@@ -9,7 +9,6 @@ import com.intellij.execution.configurations.ConfigurationFactory
 import com.intellij.execution.configurations.RefactoringListenerProvider
 import com.intellij.execution.configurations.RuntimeConfigurationError
 import com.intellij.execution.runners.ExecutionEnvironment
-import com.intellij.openapi.options.SettingsEditor
 import com.intellij.openapi.options.ShowSettingsUtil
 import com.intellij.openapi.project.Project
 import com.intellij.psi.NavigatablePsiElement
@@ -18,28 +17,20 @@ import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.refactoring.listeners.RefactoringElementAdapter
 import com.intellij.refactoring.listeners.RefactoringElementListener
 import software.amazon.awssdk.services.lambda.model.Runtime
-import software.aws.toolkits.core.credentials.CredentialProviderNotFound
 import software.aws.toolkits.core.credentials.ToolkitCredentialsProvider
 import software.aws.toolkits.core.region.AwsRegion
-import software.aws.toolkits.jetbrains.core.credentials.CredentialManager
-import software.aws.toolkits.jetbrains.core.credentials.ProjectAccountSettingsManager
-import software.aws.toolkits.jetbrains.core.region.AwsRegionProvider
-import software.aws.toolkits.jetbrains.services.lambda.HandlerCompletionProvider
 import software.aws.toolkits.jetbrains.services.lambda.Lambda.findPsiElementsForHandler
-import software.aws.toolkits.jetbrains.services.lambda.LambdaBuilder
 import software.aws.toolkits.jetbrains.services.lambda.LambdaHandlerResolver
 import software.aws.toolkits.jetbrains.services.lambda.RuntimeGroup
 import software.aws.toolkits.jetbrains.services.lambda.execution.LambdaRunConfiguration
 import software.aws.toolkits.jetbrains.services.lambda.execution.LambdaRunConfigurationBase
-import software.aws.toolkits.jetbrains.services.lambda.execution.sam.SamTemplateUtils.findFunctionsFromTemplate
+import software.aws.toolkits.jetbrains.services.lambda.sam.SamTemplateUtils.findFunctionsFromTemplate
 import software.aws.toolkits.jetbrains.services.lambda.runtimeGroup
 import software.aws.toolkits.jetbrains.services.lambda.validOrNull
 import software.aws.toolkits.jetbrains.settings.AwsSettingsConfigurable
 import software.aws.toolkits.jetbrains.settings.SamSettings
-import software.aws.toolkits.jetbrains.utils.ui.selected
 import software.aws.toolkits.resources.message
 import java.io.File
-import javax.swing.JComponent
 
 class SamRunConfigurationFactory(configuration: LambdaRunConfiguration) : ConfigurationFactory(configuration) {
     override fun createTemplateConfiguration(project: Project) = SamRunConfiguration(project, this)
@@ -54,7 +45,7 @@ class SamRunConfiguration(project: Project, factory: ConfigurationFactory) :
     RefactoringListenerProvider {
     override fun getOptions() = super.getOptions() as LocalLambdaOptions
 
-    override fun getConfigurationEditor() = SamRunSettingsEditor(project)
+    override fun getConfigurationEditor() = LocalLambdaRunSettingsEditor(project)
 
     override fun checkConfiguration() {
         if (SamSettings.getInstance().executablePath.isNullOrEmpty()) {
@@ -219,83 +210,6 @@ class SamRunConfiguration(project: Project, factory: ConfigurationFactory) :
         }
     } catch (e: Exception) {
         null
-    }
-}
-
-class SamRunSettingsEditor(project: Project) : SettingsEditor<SamRunConfiguration>() {
-    private val view = SamRunSettingsEditorPanel(project, HandlerCompletionProvider(project))
-    private val regionProvider = AwsRegionProvider.getInstance()
-    private val credentialManager = CredentialManager.getInstance()
-
-    init {
-        val supported = LambdaBuilder.supportedRuntimeGroups
-            .flatMap { it.runtimes }
-            .sorted()
-
-        val selected = RuntimeGroup.determineRuntime(project)?.let { if (it in supported) it else null }
-        val accountSettingsManager = ProjectAccountSettingsManager.getInstance(project)
-
-        view.setRuntimes(supported)
-        view.runtime.selectedItem = selected
-
-        view.regionSelector.setRegions(regionProvider.regions().values.toMutableList())
-        view.regionSelector.selectedRegion = accountSettingsManager.activeRegion
-
-        view.credentialSelector.setCredentialsProviders(credentialManager.getCredentialProviders())
-        if (accountSettingsManager.hasActiveCredentials()) {
-            view.credentialSelector.setSelectedCredentialsProvider(accountSettingsManager.activeCredentialProvider)
-        }
-    }
-
-    override fun createEditor(): JComponent = view.panel
-
-    override fun resetEditorFrom(configuration: SamRunConfiguration) {
-        view.useTemplate.isSelected = configuration.isUsingTemplate()
-        if (configuration.isUsingTemplate()) {
-            view.runtime.isEnabled = false
-            view.setTemplateFile(configuration.templateFile())
-            view.selectFunction(configuration.logicalId())
-        } else {
-            view.setTemplateFile(null) // Also clears the functions selector
-            view.runtime.model.selectedItem = configuration.runtime()
-            view.handler.setText(configuration.handler())
-        }
-
-        view.environmentVariables.envVars = configuration.environmentVariables()
-        view.regionSelector.selectedRegion = regionProvider.lookupRegionById(configuration.regionId())
-
-        configuration.credentialProviderId()?.let {
-            try {
-                view.credentialSelector.setSelectedCredentialsProvider(credentialManager.getCredentialProvider(it))
-            } catch (e: CredentialProviderNotFound) {
-                // Use the raw string here to not munge what the customer had, will also allow it to show the error
-                // that it could not be found
-                view.credentialSelector.setSelectedInvalidCredentialsProvider(it)
-            }
-        }
-
-        if (configuration.isUsingInputFile()) {
-            view.lambdaInput.inputFile = configuration.inputSource()
-        } else {
-            view.lambdaInput.inputText = configuration.inputSource()
-        }
-    }
-
-    override fun applyEditorTo(configuration: SamRunConfiguration) {
-        if (view.useTemplate.isSelected) {
-            configuration.useTemplate(view.templateFile.text, view.function.selected()?.logicalName)
-        } else {
-            configuration.useHandler(view.runtime.selected(), view.handler.text)
-        }
-
-        configuration.environmentVariables(view.environmentVariables.envVars)
-        configuration.regionId(view.regionSelector.selectedRegion?.id)
-        configuration.credentialProviderId(view.credentialSelector.getSelectedCredentialsProvider())
-        if (view.lambdaInput.isUsingFile) {
-            configuration.useInputFile(view.lambdaInput.inputFile)
-        } else {
-            configuration.useInputText(view.lambdaInput.inputText)
-        }
     }
 }
