+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/lambda/python/PythonLambdaBuilder.kt
@@ -12,7 +12,7 @@ import com.intellij.psi.PsiElement
 import software.amazon.awssdk.services.lambda.model.Runtime
 import software.aws.toolkits.jetbrains.services.lambda.BuiltLambda
 import software.aws.toolkits.jetbrains.services.lambda.LambdaBuilder
-import software.aws.toolkits.jetbrains.services.lambda.execution.sam.SamTemplateUtils
+import software.aws.toolkits.jetbrains.services.lambda.sam.SamTemplateUtils
 import java.util.concurrent.CompletionStage
 
 class PythonLambdaBuilder : LambdaBuilder() {
